Move pages into feature-specific directories
