+++ frontend/src/features/noteTaking/notebooks/pages/NotebooksPage/components/CreateNotebookButton.tsx
