+++ frontend/src/features/noteTaking/notebooks/pages/NotebooksPage/components/DeleteNotebookAlertDialog.tsx
