+++ frontend/src/features/noteTaking/notebooks/pages/NotebooksPage/components/NotebookCard.tsx
