+++ frontend/src/features/noteTaking/notebooks/pages/NotebooksPage/components/NotebookDropdownMenu.tsx
