+++ frontend/src/features/noteTaking/notebooks/pages/NotebooksPage/components/NotebookGrid.tsx
